@@ -55,6 +55,7 @@ import { ConfigDatabase, CommandLineAction, injectConfigDatabase, AcceptsConfigD
 import * as SupportsClipboardPaste from "./SupportsClipboardPaste";
 import * as SupportsDialogStack from "./SupportsDialogStack";
 import { ExtensionManager } from './extension/InternalTypes';
+import { DeepReadonlyObject, DeepReadonly } from 'extraterm-readonly-toolbox';
 
 type VirtualScrollable = VirtualScrollArea.VirtualScrollable;
 type VirtualScrollArea = VirtualScrollArea.VirtualScrollArea;
@@ -296,6 +297,7 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
       scrollArea.addEventListener('wheel', this._handleMouseWheel.bind(this), true);
       scrollContainer.addEventListener('mousedown', this._handleMouseDown.bind(this), true);
       scrollArea.addEventListener('keydown', this._handleKeyDownCapture.bind(this), true);
+      scrollArea.addEventListener("keypress", (ev) => this._handleKeyPressCapture(ev), true);
 
       scrollArea.addEventListener(VirtualScrollArea.EVENT_RESIZE, this._handleVirtualScrollableResize.bind(this));
       scrollArea.addEventListener(TerminalViewer.EVENT_KEYBOARD_ACTIVITY, () => {
@@ -409,7 +411,6 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
 
     pty.onData((text: string): void => {
       this._emulator.write(text);
-      // FIXME flow control.
     });
 
     doLater(() => {
@@ -437,10 +438,10 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
     this._extensionManager = extensionManager;
   }
 
-  private _isNoFrameCommand(commandLine: string): boolean {
+  private _commandNeedsFrame(commandLine: string): boolean {
     const cleanCommandLine = commandLine.trim();
     if (cleanCommandLine === "") {
-      return true;
+      return false;
     }
     
     const commandParts = cleanCommandLine.split(/\s+/);
@@ -448,27 +449,38 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
       return false;
     } else {
     
-      const commandLineActions = this._configManager.getConfig(COMMAND_LINE_ACTIONS_CONFIG) || [];
-      return commandLineActions
-        .filter( cla => ! cla.frame)
-        .some( cla => {
-          if (cla.matchType === 'name') {
-            const matcherParts = cla.match.split(/\s+/);
-            for (let i=0; i < matcherParts.length; i++) {
-              if (i >= commandParts.length) {
-                return false;
-              }
-              if (matcherParts[i] !== commandParts[i]) {
-                return false;
-              }
-            }
-            return true;        
-          } else {
-            // regexp
-            return (new RegExp(cla.match)).test(cleanCommandLine);
-          }
-        } );
+      const commandLineActions: DeepReadonly<CommandLineAction[]> = 
+        this._configManager.getConfig(COMMAND_LINE_ACTIONS_CONFIG) || [];
+      const frameByDefault = this._configManager.getConfig(GENERAL_CONFIG).frameByDefault;
+
+      for (const cla of commandLineActions) {
+        if (this._commandLineActionMatches(commandLine, cla)) {
+          return cla.frame;
+        }
+      }
+      return frameByDefault;
     }
+  }
+
+  private _commandLineActionMatches(command: string, cla: DeepReadonly<CommandLineAction>): boolean {
+    const cleanCommandLine = command.trim();
+    const commandParts = command.trim().split(/\s+/);
+
+    if (cla.matchType === 'name') {
+      const matcherParts = cla.match.split(/\s+/);
+      for (let i=0; i < matcherParts.length; i++) {
+        if (i >= commandParts.length) {
+          return false;
+        }
+        if (matcherParts[i] !== commandParts[i]) {
+          return false;
+        }
+      }
+      return true;        
+    } else {
+      // regexp
+      return (new RegExp(cla.match)).test(cleanCommandLine);
+    }   
   }
 
   /**
@@ -1339,7 +1351,24 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
       ev.preventDefault();
     }
   }
-  
+
+  private _handleKeyPressCapture(ev :KeyboardEvent): void {
+    if (this._terminalViewer === null || this._keyBindingManager === null ||
+        this._keyBindingManager.getKeyBindingsContexts() === null) {
+      return;
+    }
+    
+    const keyBindings = this._keyBindingManager.getKeyBindingsContexts().context(this._mode === Mode.DEFAULT
+        ? KEYBINDINGS_DEFAULT_MODE : KEYBINDINGS_CURSOR_MODE);
+    const command = keyBindings.mapEventToCommand(ev);
+    if (command != null) {
+      // We merely have to detech the key press as belonging to one of our shortcuts and then prevent
+      // it from reaching the layers below such as the terminal viewer and term emulation.
+      ev.stopPropagation();
+      ev.preventDefault();
+    }
+  }
+
   private _handleContextMenu(): void {
     if (this._terminalViewer !== null) {
       this._terminalViewer.executeCommand(COMMAND_OPEN_COMMAND_PALETTE);
@@ -1720,7 +1749,7 @@ export class EtTerminal extends ThemeableElementBase implements Commandable, Acc
       cleancommand = trimmed.slice(trimmed.indexOf(" ")).trim();
     }
     
-    if ( ! this._isNoFrameCommand(cleancommand)) {
+    if (this._commandNeedsFrame(cleancommand)) {
       // Create and set up a new command-frame.
       const el = this._createEmbeddedViewerElement();
 
