@@ -7,7 +7,7 @@ import * as _ from 'lodash';
 import Vue from 'vue';
 
 import { FrameSettingsUi, nextId, Identifiable, IdentifiableCommandLineAction} from './FrameSettingsUi';
-import { COMMAND_LINE_ACTIONS_CONFIG, ConfigKey, CommandLineAction } from '../../Config';
+import { COMMAND_LINE_ACTIONS_CONFIG, ConfigKey, CommandLineAction, GENERAL_CONFIG, GeneralConfig } from '../../Config';
 import { Logger, getLogger } from '../../logging/Logger';
 import log from '../../logging/LogDecorator';
 import { SettingsBase } from './SettingsBase';
@@ -19,13 +19,13 @@ export class FrameSettings extends SettingsBase<FrameSettingsUi> {
   private _log: Logger = null;
 
   constructor() {
-    super(FrameSettingsUi, [COMMAND_LINE_ACTIONS_CONFIG]);
+    super(FrameSettingsUi, [GENERAL_CONFIG, COMMAND_LINE_ACTIONS_CONFIG]);
     this._log = getLogger(FRAME_SETTINGS_TAG, this);
   }
 
   protected _setConfig(key: ConfigKey, config: any): void {
+    const ui = this._getUi();
     if (key === COMMAND_LINE_ACTIONS_CONFIG) {
-      const ui = this._getUi();
       const commandLineActions = <CommandLineAction[]> config;
 
       const cleanCommandLineAction = _.cloneDeep(ui.commandLineActions);
@@ -37,6 +37,11 @@ export class FrameSettings extends SettingsBase<FrameSettingsUi> {
         ui.commandLineActions = updateCLA;
       }
     }
+
+    if (key === GENERAL_CONFIG) {
+      const generalConfig = <GeneralConfig> config;
+      ui.frameByDefault = config.frameByDefault ? "true" : "false";
+    }
   }
 
   protected _dataChanged(): void {
@@ -44,6 +49,10 @@ export class FrameSettings extends SettingsBase<FrameSettingsUi> {
     const commandLineActions = _.cloneDeep(ui.commandLineActions);
     stripIds(commandLineActions);
     this._updateConfig(COMMAND_LINE_ACTIONS_CONFIG, commandLineActions);
+
+    const generalConfig = <GeneralConfig> this._getConfigCopy(GENERAL_CONFIG);
+    generalConfig.frameByDefault = ui.frameByDefault === "true" ? true : false;
+    this._updateConfig(GENERAL_CONFIG, generalConfig);
   }
 }
 
